log-service/requirements.txt
fastapi
uvicorn
pymongo
motor
aiogram
pydantic
cloudinary
jinja2
requests
python-multipart
python-decouple==3.8
pymongo